% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraints_to_query}
\alias{constraints_to_query}
\title{Build a query string from a Constraints object}
\usage{
constraints_to_query(constraints)
}
\arguments{
\item{constraints}{A \code{\link[=Constraints]{Constraints()}} object.}
}
\value{
Query string without leading \code{?}.
}
\description{
Build a query string from a Constraints object
}
