% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dab.r-package.R
\docType{package}
\name{dab.r-package}
\alias{dab.r}
\alias{dab.r-package}
\title{dab.r: DAB Bindings for R}
\description{
R client for GeoDAB Data Access Bridge (DAB) services, including the Terms API and OGC OM-JSON API (WHOS and HIS-Central). Mirrors the capabilities of the dab-py Python package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ESSI-Lab/dab-r}
  \item Report bugs at \url{https://github.com/ESSI-Lab/dab-r/issues}
}

}
\author{
\strong{Maintainer}: ESSI-Lab \email{info@essi-lab.eu} [copyright holder, funder]

Authors:
\itemize{
  \item Enrico Boldrini (Python reference implementation)
}

}
\keyword{internal}
