% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{download_constraints_to_query}
\alias{download_constraints_to_query}
\title{Build a query string from a DownloadConstraints object}
\usage{
download_constraints_to_query(download_constraints)
}
\arguments{
\item{download_constraints}{A \code{\link[=DownloadConstraints]{DownloadConstraints()}} object.}
}
\value{
Query string without leading \code{?}.
}
\description{
Build a query string from a DownloadConstraints object
}
