% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om-api-models.R
\name{downloads_to_df}
\alias{downloads_to_df}
\title{Convert downloads to a data frame}
\usage{
downloads_to_df(collection)
}
\arguments{
\item{collection}{Object returned by \code{client$get_download_status()}.}
}
\value{
A data frame.
}
\description{
Convert downloads to a data frame
}
