% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om-api-models.R
\name{features_to_df}
\alias{features_to_df}
\title{Convert features to a data frame (current page)}
\usage{
features_to_df(collection)
}
\arguments{
\item{collection}{Object returned by \code{client$get_features()}.}
}
\value{
A data frame.
}
\description{
Convert features to a data frame (current page)
}
