% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om-api-client.R
\name{HISCentralClient}
\alias{HISCentralClient}
\title{Create a HIS-Central OM-API client}
\usage{
HISCentralClient(token, view = "his-central")
}
\arguments{
\item{token}{GeoDAB token embedded in the URL path.}

\item{view}{GeoDAB view (e.g. \code{"whos"}, \code{"his-central"}).}
}
\value{
A \link{DABClient} R6 object configured for HIS-Central.
}
\description{
Create a HIS-Central OM-API client
}
