% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om-api-models.R
\name{observations_all_to_df}
\alias{observations_all_to_df}
\title{Convert all paginated observations to a data frame}
\usage{
observations_all_to_df(collection)
}
\arguments{
\item{collection}{Object returned by \code{client$get_observations()}. Fetch
further pages with \code{collection$next_page()} before calling this
function.}
}
\value{
A data frame of every observation accumulated in the collection.
}
\description{
Convert all paginated observations to a data frame
}
