% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om-api-models.R
\name{points_to_df}
\alias{points_to_df}
\title{Convert observation points to a data frame}
\usage{
points_to_df(observation)
}
\arguments{
\item{observation}{Object returned by \code{client$get_observation_with_data()}.}
}
\value{
A data frame with columns \code{Time} and \code{Value}.
}
\description{
Convert observation points to a data frame
}
