% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PREDEFINED_SEARCH_AREA}
\alias{PREDEFINED_SEARCH_AREA}
\title{Predefined search areas property for \code{get_properties()}}
\format{
An object of class \code{character} of length 1.
}
\usage{
PREDEFINED_SEARCH_AREA
}
\description{
Pass to \code{client$get_properties(PREDEFINED_SEARCH_AREA)} and as
\code{predefinedSearchArea} in \code{\link[=Constraints]{Constraints()}}.
}
\keyword{datasets}
