% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om-api-properties.R
\name{properties_to_df}
\alias{properties_to_df}
\title{Convert a properties result to a data frame}
\usage{
properties_to_df(properties)
}
\arguments{
\item{properties}{Object returned by \code{client$get_properties()}.}
}
\value{
A data frame with columns Label, Value, and Observation Count.
}
\description{
Convert a properties result to a data frame
}
