% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms-api.R
\name{TermsAPI}
\alias{TermsAPI}
\title{Create a DAB Terms API client}
\usage{
TermsAPI(token, view)
}
\arguments{
\item{token}{GeoDAB token.}

\item{view}{GeoDAB view (e.g. \code{"blue-cloud-terms"}).}
}
\value{
A \link{TermsAPI} R6 object.
}
\description{
Create a DAB Terms API client
}
\examples{
\dontrun{
api <- TermsAPI(token = "my-token", view = "blue-cloud-terms")
terms <- api$get_terms(type = "instrument", max = 10)
}
}
