% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om-api-models.R
\name{observations_to_df}
\alias{observations_to_df}
\title{Convert observations to a data frame (current page)}
\usage{
observations_to_df(collection)
}
\arguments{
\item{collection}{Object returned by \code{client$get_observations()}.}
}
\value{
A data frame for \code{collection$current_page_obs} only.
}
\description{
Convert observations to a data frame (current page)
}
\seealso{
\code{\link[=observations_all_to_df]{observations_all_to_df()}} for all pages after \code{$next_page()}.
}
